Funflow.Config

Documentation

type ConfigKey = Text

type ConfigMap = Object

type EnvConfigMap = HashMap Text String

data ExternalConfig

Constructors

ExternalConfig

Fields

fileConfig :: ConfigMap

envConfig :: EnvConfigMap

data Configurable a where

A value which is intended to be populated using an external source (e.g. a config file)

Constructors

ConfigFromFile :: FromJSON a => ConfigKey -> Configurable a
Define a configurable which will be loaded from a config file using the given key

ConfigFromEnv :: FromJSON a => ConfigKey -> Configurable a
Define a configurable which will be loaded from the specified environment variable

Literal :: a -> Configurable a
A literal value which does not need to be loaded from an external config source

render :: forall a. Configurable a -> ExternalConfig -> Either String (Configurable a)

Render a Configurable into a Literal value using a set of external configurations, returning an error message if rendering failed.

getConfigKey :: Configurable a -> Maybe ConfigKey

Gets the config key for a configurable value, if it exists.

data ConfigKeysBySource

Stores ConfigKey values by their declared sources.

Constructors

ConfigKeysBySource

Fields

fileConfigKeys :: HashSet Text

envConfigKeys :: HashSet Text

configKeyBySource :: Configurable a -> ConfigKeysBySource

Get the key of a Configurable as a ConfigKeysBySource.

missing :: ExternalConfig -> ConfigKeysBySource -> [ConfigKey]

Get a list of any ConfigKeys which don't exist in their corresponding field in the providedExternalConfig

readEnv :: MonadIO m => ConfigKey -> m (HashMap Text String)

Construct an HashMap containing specified environment variable values.

readEnvs :: MonadIO m => [ConfigKey] -> m (HashMap Text String)

Convenience function for calling readEnv on a list of ConfigKeys

readYamlFileConfig :: (MonadIO m, FromJSON a) => FilePath -> m a

Construct a HashMap containing the content of a yaml file. Is just an Alias for decodeFileThrow.