Funflow.Effects.Docker

Description

Run commands using Docker

Documentation

data DockerEffectConfig

Configure what task to run in Docker

Constructors

DockerEffectConfig

Fields

image :: Text

The name of the docker image

command :: Text

The command to run

args :: [Arg]

The arguments to pass to the command run inside of the container

data DockerEffect i o where

Constructors

DockerEffect :: DockerEffectConfig -> DockerEffect DockerEffectInput Item

data DockerEffectInput

Input to a Docker effect to finalize its configuration

Constructors

DockerEffectInput

Fields

inputBindings :: [VolumeBinding]

Input items to mount on the container

argsVals :: Map String Text

A map representing how to fill the argument placeholders (placeholder label -> argument value)

data VolumeBinding

Represent how to bind a directory from cas-store (CS.Item) to a container internal file system

Constructors

VolumeBinding

Fields

item :: Item

mount :: Path Abs Dir

data Arg

Represent an argument to pass to the command run inside of a Docker container

Constructors

Arg Text
Raw text argument

Placeholder String
A placeholder for an argument to be passed as runtime input to the effect (filled by argsVals)