Funflow.Flow

Documentation

type Flow input output = forall m. MonadIO m => AnyRopeWith RequiredStrands (RequiredCoreEffects m) input output

Flow is the main type of Funflow. It is a task that takes an input value of type input and produces an output value of type output. It can use any named effect (strand) that is defined in RequiredStrands.

type RequiredStrands = '['("simple", SimpleEffect), '("store", StoreEffect), '("docker", DockerEffect)]

The constraints on the set of "strands" These will be "interpreted" into "core effects" (which have contraints defined below).

type RequiredCoreEffects m = '[Arrow, ArrowChoice, HasKleisli m, ProvidesCaching]

The class constraints on the "core effect". The "core effect" is the effect used to run any kind of "binary effect" ("strand")

toFlow :: IsFlow binEff => binEff i o -> Flow i o

pureFlow :: (i -> o) -> Flow i o

ioFlow :: (i -> IO o) -> Flow i o

dockerFlow :: DockerEffectConfig -> Flow DockerEffectInput Item

putDir :: Flow (Path Abs Dir) Item

getDir :: Flow Item (Path Abs Dir)