Funflow.Run

Description

This module defines how to run your flows

Documentation

runFlow

Arguments

:: LooseRopeWith RequiredStrands (RequiredCore IO) input output
The flow to run

-> input
The input to evaluate the flow with

-> IO output

Run a flow with the default configuration

runFlowWithConfig

Arguments

:: RunFlowConfig
The configuration of the flow

-> LooseRopeWith RequiredStrands (RequiredCore IO) input output
The flow to run

-> input
The input to evaluate the flow with

-> IO output

Run a flow, parsing any required Configurable values from their respective sources. This flow executor includes interpreters for the following tasks:

SimpleTask

StoreTask

DockerTask - The container is run with working directory '/workdir'. Files written to this directory are included in the tasks's Item output.

data RunFlowConfig

Flow execution configuration

Constructors

RunFlowConfig

Fields

storePath :: Path Abs Dir

configFile :: Maybe (Path Abs File)