Funflow.Tasks.Docker

Description

Run commands using Docker

Documentation

data DockerTaskConfig

Configure what task to run in Docker

Constructors

DockerTaskConfig

Fields

image :: Text

The name of the docker image

command :: Text

The command to run

args :: [Arg]

The arguments to pass to the command run inside of the container

data DockerTask i o where

Constructors

DockerTask :: DockerTaskConfig -> DockerTask DockerTaskInput Item

data DockerTaskInput

Input to a Docker task to finalize its configuration

Constructors

DockerTaskInput

Fields

inputBindings :: [VolumeBinding]

Input items to mount on the container

argsVals :: Map String Text

A map representing how to fill the argument placeholders (placeholder label -> argument value)

data VolumeBinding

Represent how to bind a directory from cas-store (CS.Item) to a container internal file system

Constructors

VolumeBinding

Fields

item :: Item

mount :: Path Abs Dir

data Arg

Represent an argument to pass to the command run inside of a Docker container

Constructors

Arg (Configurable Text)
Raw text argument

Placeholder String
A placeholder for an argument to be passed as runtime input to the task (filled by argsVals)

renderArg :: ExternalConfig -> Arg -> Either String Arg

Renders an Arg with external configurations

getIdFromArg :: Arg -> Maybe ConfigKeysBySource

Extracts a ConfigKey from an Arg, if it exists.